Extract ObjectStorage to separate file
